Reorganize codebase, add E2E CLI tests, docs and security updates
Split large files for better navigability:
- task.go → task.go (mutations) + query.go (reads)
- id.go merged into store.go
- task_cmds.go → task_cmds.go + list_cmd.go + query_cmds.go
- config commands extracted to config_cmd.go
- Added doc.go with package layout guide

Add E2E CLI test suite (21 tests) that builds the binary and
exercises every major command. Caught and fixed a bug where
--actor flag was not wired to create command's CreatedBy field.

Add SECURITY.md, update docs with config set, ExcludeStatuses,
tree view, web UI no-auth note, and testing/code quality guides.
