+++ .github/workflows/test.yml
@@ -17,5 +17,8 @@ jobs:
       - name: Build
         run: go build ./...
 
-      - name: Test
+      - name: Unit Tests
         run: go test ./... -v -count=1
+
+      - name: E2E CLI Tests
+        run: go test -tags=e2e -v -count=1 ./cmd/gig/
