+++ store.go
@@ -1,15 +1,38 @@
 package gig
 
 import (
+	"crypto/sha256"
 	"database/sql"
+	"encoding/hex"
 	"fmt"
 	"os"
 	"path/filepath"
 	"sync"
+	"time"
 
+	"github.com/google/uuid"
 	"github.com/neerajg/gig/internal/migrate"
 	_ "modernc.org/sqlite"
 )
+
+const defaultHashLength = 4
+
+// GenerateID produces a short, prefix-based ID like "gig-a3f8".
+// The hash is derived from a UUID + current timestamp to ensure uniqueness.
+func GenerateID(prefix string, hashLen int) string {
+	if prefix == "" {
+		prefix = "gig"
+	}
+	if hashLen < 3 || hashLen > 8 {
+		hashLen = defaultHashLength
+	}
+
+	raw := fmt.Sprintf("%s-%d", uuid.New().String(), time.Now().UnixNano())
+	sum := sha256.Sum256([]byte(raw))
+	hash := hex.EncodeToString(sum[:])[:hashLen]
+
+	return fmt.Sprintf("%s-%s", prefix, hash)
+}
 
 // Store is the main entry point for the gig SDK.
 // It holds the database connection, configuration, and event listeners.
